+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/iteractionapi/client/PaymentClient.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.iteractionapi.dto.order.OrderDto;
 import ru.yandex.practicum.iteractionapi.dto.payment.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @FeignClient(name = "payment", configuration = FeignConfig.class)
@@ -14,13 +15,13 @@ public interface PaymentClient {
     PaymentDto createPayment(@RequestBody OrderDto orderDto);
 
     @PostMapping("/api/v1/payment/totalCost")
-    Double calculateTotalCost(@RequestBody OrderDto orderDto);
+    BigDecimal calculateTotalCost(@RequestBody OrderDto orderDto);
 
     @PostMapping("/api/v1/payment/refund")
     boolean successPayment(@RequestBody UUID orderId);
 
     @PostMapping("/api/v1/payment/productCost")
-    Double calculateProductCost(@RequestBody OrderDto orderDto);
+    BigDecimal calculateProductCost(@RequestBody OrderDto orderDto);
 
     @PostMapping("/api/v1/payment/failed")
     boolean failedPayment(@RequestBody UUID orderId);
