+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/iteractionapi/dto/order/OrderDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.UUID;
 
@@ -22,7 +23,7 @@ public class OrderDto {
     private Double deliveryWeight;
     private Double deliveryVolume;
     private Boolean fragile;
-    private Double totalPrice;
-    private Double deliveryPrice;
-    private Double productPrice;
+    private BigDecimal totalPrice;
+    private BigDecimal deliveryPrice;
+    private BigDecimal productPrice;
 }
