+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/iteractionapi/dto/payment/PaymentDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -13,7 +14,7 @@ import java.util.UUID;
 @Builder
 public class PaymentDto {
     private UUID paymentId;
-    private Double totalPayment;
-    private Double deliveryTotal;
-    private Double feeTotal;
+    private BigDecimal totalPayment;
+    private BigDecimal deliveryTotal;
+    private BigDecimal feeTotal;
 }
