+++ commerce/payment/src/main/java/ru/yandex/practicum/payment/service/PaymentService.java
@@ -3,12 +3,13 @@ package ru.yandex.practicum.payment.service;
 import ru.yandex.practicum.iteractionapi.dto.order.OrderDto;
 import ru.yandex.practicum.iteractionapi.dto.payment.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
     PaymentDto createPayment(OrderDto orderDto);
-    Double calculateTotalCost(OrderDto orderDto);
+    BigDecimal calculateTotalCost(OrderDto orderDto);
     void successPayment(UUID orderId);
-    Double calculateProductCost(OrderDto orderDto);
+    BigDecimal calculateProductCost(OrderDto orderDto);
     void failedPayment(UUID orderId);
 }
